% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnet.r
\name{as_igraph}
\alias{as_igraph}
\title{Coerce objects \code{rn_edgelist} to \code{igraph}}
\usage{
as_igraph(x, ...)
}
\arguments{
\item{x}{An object of class \code{rn_edgelist} or a list of that.}

\item{...}{Further arguments passed to \code{\link[igraph:graph_from_data_frame]{graph_from_data_frame}}
from the \pkg{igraph} R package.}
}
\description{
Coerce objects \code{rn_edgelist} to \code{igraph}
}
