% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnet.r
\name{as_rn_edgelist}
\alias{as_rn_edgelist}
\title{Coerce different graph objects to \code{rn_edgelist}}
\usage{
as_rn_edgelist(x)
}
\arguments{
\item{x}{An object to be coerced to \link{rn_edgelist}}
}
\description{
Coerce different graph objects to \code{rn_edgelist}
}
