% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_together.r
\name{edgelist_to_adjmat_w_attributes}
\alias{edgelist_to_adjmat_w_attributes}
\title{Creates an adjacency matrix from an edgelist with vertex attributes}
\usage{
edgelist_to_adjmat_w_attributes(edgelist, data,
  data.idvar = colnames(data)[1], ids.to.keep = NULL, data.timevar = NULL,
  edgelist.timevar = NULL)
}
\arguments{
\item{edgelist}{A data.frame. Only the first two columns will be used.}

\item{data}{A data.frame with vertex attributes.}

\item{data.idvar}{Character scalar. Name of the variable that holds the
vertex id (label). By default it uses the first column.}
}
\value{
A list with the following
\item{adjmat}{A matrix of \eqn{n\times n}.}
\item{data}{A sorted version of \code{data} according to the network coding.}
}
\description{
The only difference with \link[igraph:graph_from_data_frame]{igraph::graph_from_data_frame} is that
this returns an 0/1 adjacency matrix and sorts whatever attributes were
passed via \code{data} accordingly to the network coding.
}
\examples{
# Simple example
net <- c("a", "b", "a", "c", "b", "c", "c", "b", "c", "a")
net <- matrix(net, byrow = TRUE, ncol=2)

dat <- data.frame(name = c("d", "a", "c", "b"), age = c(4, 1, 3, 2))
ans0 <- edgelist_to_adjmat_w_attributes(net, dat)
}
