% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnet.r
\name{encode_mat}
\alias{encode_mat}
\title{This function takes a two-column matrix, generates factors and encodeds it
into an integer matrix with an extra attribute: \code{labels}}
\usage{
encode_mat(x)
}
\description{
This function takes a two-column matrix, generates factors and encodeds it
into an integer matrix with an extra attribute: \code{labels}
}
