% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnet.r
\name{makegroups}
\alias{makegroups}
\title{This function takes a set of vectors and creates a unique id using all
by pasting it.}
\usage{
makegroups(dat, group.id, sep)
}
\description{
This function takes a set of vectors and creates a unique id using all
by pasting it.
}
