% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnet.r
\name{rn_edgelist_to_igraph}
\alias{rn_edgelist_to_igraph}
\title{Turn a}
\usage{
rn_edgelist_to_igraph(x, ...)
}
\arguments{
\item{x}{An object of class \code{rn_edgelist} or a list of that.}

\item{...}{Further arguments passed to \href{igraph:graph_from_data_frame}{graph_from_data_frame}}
}
\description{
Turn a
}
