% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnet.r
\name{set_as_char}
\alias{set_as_char}
\title{This function coerces a set of variables in a data frame into character
so that when that is turned into a factor, information is not loss.}
\usage{
set_as_char(dat, vars)
}
\description{
This function coerces a set of variables in a data frame into character
so that when that is turned into a factor, information is not loss.
}
