% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnet.r
\name{survey_to_edgelist}
\alias{survey_to_edgelist}
\title{Creates an edgelist (or a list of edgelist) from survey data}
\usage{
survey_to_edgelist(dat, ego.id, alter.id, group.id = NULL, time = NULL,
  group.sep = "-", time.sep = "-")
}
\arguments{
\item{dat}{\code{data.frame}.}

\item{ego.id}{Character scalar.}

\item{alter.id}{Character vector.}

\item{group.id}{Character vector.}

\item{time}{Character vector.}

\item{group.sep}{Character scalar.}

\item{time.sep}{Character scalar.}
}
\value{
If no \code{group.id} or \code{time} are provided, an integer matrix with
two named columns (\emph{ego} and \emph{alter}), and an attribute "\code{labels}". Otherwise
will return a nested list in which the first level is given by \code{group.id},
and the second level by \code{time} (if provided).
}
\description{
The function grabs ego ids and its nominations and creates an edgelist
in which the labels (ids) are recoded such that ids go from 1 to n.
}
\examples{

1+1
}
